feat: implement benchmarking methods for performance comparison between traditional approach and Superfilter
