+++ Benchmark/UltraLargeScaleTest.cs
@@ -1,0 +1,41 @@
+using BenchmarkDotNet.Attributes;
+using BenchmarkDotNet.Jobs;
+using Superfilter.Constants;
+using Superfilter.Defaults;
+using Superfilter.Entities;
+
+[MemoryDiagnoser]
+[SimpleJob(RuntimeMoniker.Net90)]
+internal class UltraLargeScaleTest(int datasetSize)
+{
+    private readonly DefaultHasFilters _defaultHasFilters = new([
+        new FilterCriterion("User.Name", Operator.Contains, "John"),
+        new FilterCriterion("User.Age", Operator.GreaterThan, "30"),
+        new FilterCriterion("User.Country", Operator.In, "USA,Canada,France")
+    ]);
+
+    private List<User> _users = [];
+
+    // Test extrême pour voir si on peut atteindre le point de basculement
+    [Params(2_000_000, 5_000_000)] private int DatasetSize { get; } = datasetSize;
+
+    [GlobalSetup]
+    public void Setup()
+    {
+        Console.WriteLine($"🚀 Génération de {DatasetSize:N0} utilisateurs (test extrême)...");
+        _users = UserGenerator.GenerateUsers(DatasetSize);
+        Console.WriteLine($"✅ Dataset extrême de {DatasetSize:N0} utilisateurs prêt");
+    }
+
+    [Benchmark]
+    public List<User> StandardApproach()
+    {
+        return FilterMethods.StandardApproachMethod(_users, _defaultHasFilters);
+    }
+
+    [Benchmark]
+    public List<User> SuperfilterApproach()
+    {
+        return FilterMethods.SuperfilterApproachMethod(_users, _defaultHasFilters);
+    }
+}
